+++ apps/e2e/ag-grid-storybook-e2e/src/e2e/data-manager.component.cy.ts
@@ -1,5 +1,14 @@
 import { E2eVariations } from '@skyux-sdk/e2e-schematics';
 
+// This is to mitigate a Cypress issue (https://github.com/cypress-io/cypress/issues/20341) where a ResizeObserver exception is thrown.
+Cypress.on(
+  'uncaught:exception',
+  (err) =>
+    !err.message.includes(
+      'ResizeObserver loop completed with undelivered notifications.'
+    )
+);
+
 describe(`ag-grid-storybook data manager`, () => {
   E2eVariations.forEachTheme((theme) => {
     describe(`in ${theme} theme`, () => {