+++ libs/components/autonumeric/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/modules/autonumeric/autonumeric-options-provider';
-export * from './lib/modules/autonumeric/autonumeric-options';
-export * from './lib/modules/autonumeric/autonumeric.module';
+export { SkyAutonumericOptionsProvider } from './lib/modules/autonumeric/autonumeric-options-provider';
+export { SkyAutonumericOptions } from './lib/modules/autonumeric/autonumeric-options';
+export { SkyAutonumericModule } from './lib/modules/autonumeric/autonumeric.module';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.