+++ libs/components/http/testing/src/public-api.ts
@@ -1,3 +1,3 @@
-export * from './auth-http-client/auth-http-client-testing.module';
-export * from './auth-http-client/auth-http-testing-controller';
-export * from './auth-http-client/auth-token-mock-provider';
+export { SkyAuthHttpClientTestingModule } from './auth-http-client/auth-http-client-testing.module';
+export { SkyAuthHttpTestingController } from './auth-http-client/auth-http-testing-controller';
+export { SkyAuthTokenMockProvider } from './auth-http-client/auth-token-mock-provider';