+++ libs/components/phone-field/testing/src/public-api.ts
@@ -1,2 +1,2 @@
-export * from './phonefield/phone-field-testing.module';
-export * from './phonefield/phone-field-fixture';
+export { SkyPhoneFieldTestingModule } from './phonefield/phone-field-testing.module';
+export { SkyPhoneFieldFixture } from './phonefield/phone-field-fixture';