+++ libs/components/split-view/testing/src/public-api.ts
@@ -1,4 +1,4 @@
-export * from './split-view/split-view-testing.module';
-export * from './split-view/split-view-fixture';
-export * from './split-view/split-view-fixture-drawer';
-export * from './split-view/split-view-fixture-workspace';
+export { SkySplitViewTestingModule } from './split-view/split-view-testing.module';
+export { SkySplitViewFixture } from './split-view/split-view-fixture';
+export { SkySplitViewFixtureDrawer } from './split-view/split-view-fixture-drawer';
+export { SkySplitViewFixtureWorkspace } from './split-view/split-view-fixture-workspace';