+++ libs/components/theme/src/index.ts
@@ -1,17 +1,17 @@
-export * from './lib/icons/icon-manifest';
-export * from './lib/icons/icon-manifest-glyph';
-export * from './lib/icons/icon-manifest.service';
-export * from './lib/style-loader';
-export * from './lib/theme.module';
-export * from './lib/theming/theme';
-export * from './lib/theming/theme-mode';
-export * from './lib/theming/theme-settings';
-export * from './lib/theming/theme-settings-change';
-export * from './lib/theming/theme-spacing';
-export * from './lib/theming/theme.service';
-export * from './lib/viewport/viewport.service';
-export * from './lib/viewport/viewport-reserve-args';
-export * from './lib/viewport/viewport-reserve-position-type';
+export { SkyThemeIconManifest } from './lib/icons/icon-manifest';
+export { SkyThemeIconManifestGlyph } from './lib/icons/icon-manifest-glyph';
+export { SkyThemeIconManifestService } from './lib/icons/icon-manifest.service';
+export { SkyAppStyleLoader } from './lib/style-loader';
+export { SkyThemeModule } from './lib/theme.module';
+export { SkyTheme } from './lib/theming/theme';
+export { SkyThemeMode } from './lib/theming/theme-mode';
+export { SkyThemeSettings } from './lib/theming/theme-settings';
+export { SkyThemeSettingsChange } from './lib/theming/theme-settings-change';
+export { SkyThemeSpacing } from './lib/theming/theme-spacing';
+export { SkyThemeService } from './lib/theming/theme.service';
+export { SkyAppViewportService } from './lib/viewport/viewport.service';
+export { SkyAppViewportReserveArgs } from './lib/viewport/viewport-reserve-args';
+export { SkyAppViewportReservedPositionType } from './lib/viewport/viewport-reserve-position-type';
 
 // Components and directives must be exported to support Angular's "partial" Ivy compiler.
 // Obscure names are used to indicate types are not part of the public API.