+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [9.0.0-alpha.2](https://github.com/blackbaud/skyux/compare/9.0.0-alpha.1...9.0.0-alpha.2) (2023-08-09)
+
+
+### ⚠ BREAKING CHANGES
+
+* **components/ag-grid:** support AG Grid 29 ([#1549](https://github.com/blackbaud/skyux/issues/1549))
+
+### Features
+
+* **components/ag-grid:** support AG Grid 29 ([#1549](https://github.com/blackbaud/skyux/issues/1549)) ([7e8dd3d](https://github.com/blackbaud/skyux/commit/7e8dd3d3d929178632e8dcfa4d004b7eb5d0fcc1))
+* **components/packages:** update ag-grid in ng update group ([#1562](https://github.com/blackbaud/skyux/issues/1562)) ([baa5bb9](https://github.com/blackbaud/skyux/commit/baa5bb98f1542cfd7b41fc7196d41ea89b658399))
+
+
+### Bug Fixes
+
+* **components/packages:** replace '@circlon/angular-tree-component' import paths ([#1560](https://github.com/blackbaud/skyux/issues/1560)) ([2d0cfa0](https://github.com/blackbaud/skyux/commit/2d0cfa0ac16954b4060dd1964b8f09da98989458))
+* replace '*' with named exports for the public API ([#1559](https://github.com/blackbaud/skyux/issues/1559)) ([ed3b604](https://github.com/blackbaud/skyux/commit/ed3b6041eebee0fc8498010f1f1d9d514d327731))
+
+## [8.7.2](https://github.com/blackbaud/skyux/compare/8.7.1...8.7.2) (2023-08-08)
+
+
+### Bug Fixes
+
+* **components/forms:** fix input/label association in input box code example ([#1550](https://github.com/blackbaud/skyux/issues/1550)) ([bf12fb1](https://github.com/blackbaud/skyux/commit/bf12fb1169d17ba24100764e26a61247c91e667a))
+
 ## [9.0.0-alpha.1](https://github.com/blackbaud/skyux/compare/9.0.0-alpha.0...9.0.0-alpha.1) (2023-08-08)